requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=5.4.0

# Commercial API clients
openai>=1.0.0
google-generativeai>=0.3.0
boto3>=1.26.0
botocore>=1.29.0

# Open source computer vision
ultralytics>=8.0.0  # YOLOv8
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.21.0
diffusers>=0.10.0

# Grounding DINO dependencies
groundingdino-py>=0.1.0
supervision>=0.6.0

# Segment Anything
segment-anything>=1.0

# BLIP-2
salesforce-lavis>=1.0.0

# Video processing
moviepy>=1.0.3
imageio>=2.19.0

# Utilities
requests>=2.28.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.15.0 